+++ ceammc/ext/src/base/load_msg.h
@@ -23,12 +23,14 @@ class LoadMsg : public BaseObject {
 
 public:
     LoadMsg(const PdArgs& args);
+
     void output();
 
     void onClick(t_floatarg xpos, t_floatarg ypos, t_floatarg shift, t_floatarg ctrl, t_floatarg alt) override;
     void onLoadBang() override;
 
     virtual void doOutput(const AtomListView& lv);
+    virtual void doSendPd(const AtomListView& lv);
 
 private:
     void realizeDollars();